requirements.txt
# FastAPI & Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# WebSocket & Async
websockets==12.0
aiohttp==3.9.1
httpx==0.25.2

# HumeAI Integration
hume==0.5.0

# Dialer Integration (Twilio/Vonage)
twilio==8.10.3
vonage==3.11.0

# Audio Processing
pydub==0.25.1
numpy==1.26.2
pyaudio==0.2.14

# Browser Automation
selenium==4.27.1
webdriver-manager==4.0.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2

# Task Queue (Optional)
celery==5.3.4
redis[hiredis]==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
